requirements.txt
certifi==2019.9.11
chardet==3.0.4
cplex==12.10.0.0
docloud==1.0.375
docplex==2.11.176
idna==2.8
msgpack==1.0.0
numpy==1.18.2
pkg-resources==0.0.0
PyYAML==5.3.1
requests==2.22.0
scipy==1.4.1
six==1.12.0
ujson==2.0.3
urllib3==1.25.8
